requirements.txt
mysql-connector-python==8.0.14
requests==2.21.0